# docs/swagger.yaml
basePath: /
definitions:
  controller.ConflictErrorResponse:
    description: Resposta retornada quando o nickname ou email já estão em uso.
    properties:
      email:
        description: Email único do usuário
        example: email already in use
        type: string
      nickname:
        description: Apelido único do usuário
        example: nickname already in use
        type: string
    type: object
...
  input.CreateUser:
    description: Dados necessários para registrar um novo usuário no sistema
    properties:
      birthdate:
        description: Data de nascimento no formato YYYY-MM-DD (opcional)
        example: "1990-01-15"
        type: string
      email:
        description: Email único do usuário
        example: joao@email.com
        type: string
      name:
        description: Nome completo do usuário
        example: João Silva
        type: string
      nickname:
        description: Apelido único do usuário (3-32 caracteres, apenas minúsculas)
        example: joaosilva
        maxLength: 32
        minLength: 3
        type: string
      password:
        description: Senha do usuário
        example: senhaSegura123
        type: string
    required:
    - email
    - name
    - nickname
    - password
    type: object
...
  input.Identifier:
    description: Identificador para localizar um usuário no sistema
    properties:
      identifier:
        description: Identificador do usuário (email ou nickname)
        example: usuario@email.com
        type: string
      identifierType:
        description: Tipo de identificador usado (email ou nickname)
        enum:
        - email
        - nickname
        example: email
        type: string
    required:
    - identifier
    - identifierType
    type: object
  input.LoginCredentials:
    description: Credenciais para autenticação do usuário
    properties:
      identifier:
        description: Identificador do usuário (email ou nickname)
        example: usuario@email.com
        type: string
      identifierType:
        description: Tipo de identificador usado (email ou nickname)
        enum:
        - email
        - nickname
        example: email
        type: string
      password:
        description: Senha do usuário
        example: senha123
        type: string
    required:
    - identifier
    - identifierType
    - password
    type: object
...
  input.ResetPassword:
    description: Dados necessários para redefinir a senha do usuário
    properties:
      newPassword:
        description: Nova senha do usuário
        example: novaSenha123
        type: string
      resetToken:
        description: Token de redefinição de senha recebido por email
        example: abc123def456
        type: string
    required:
    - newPassword
    - resetToken
    type: object
...
  response.AuthTokensResponse:
    description: Tokens de autenticação JWT
    properties:
      accessToken:
        description: Token de acesso JWT (válido por tempo limitado)
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refreshToken:
        description: Token de refresh JWT (usado para obter novos tokens)
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  response.GetCityReponse:
    description: Informações de uma cidade retornada na busca/listagem
    properties:
      cityId:
        description: ID da cidade
        example: 1
        type: integer
      cityName:
        description: Nome da cidade
        example: São Paulo
        type: string
      countryName:
        description: Nome do país
        example: Brasil
        type: string
      stateName:
        description: Nome do estado
        example: SP
        type: string
    type: object
...
  response.UserCreatedResponse:
    description: Resposta retornada quando um usuário é criado com sucesso
    properties:
      birthdate:
        description: Data de nascimento (se houver)
        type: string
      createdAt:
        description: Data de criação
        type: string
      email:
        description: Email do usuário
        type: string
      name:
        description: Nome completo do usuário
        type: string
      nickname:
        description: Apelido único do usuário
        type: string
      profilePic:
        description: URL da foto de perfil (se houver)
        type: string
      updatedAt:
        description: Data de última atualização
        type: string
    type: object
...
  src_core_auth_controller.ErrorMessageResponse:
    description: Resposta retornada quando ocorre um erro que não requer detalhes
      adicionais, apenas uma mensagem descritiva.
    properties:
      message:
        description: Mensagem descritiva do erro
        type: string
    type: object
  src_core_auth_controller.ValidationErrorResponse:
    description: Resposta retornada quando há erros de validação nos dados enviados.
      O campo "errors" contém um mapa onde a chave é o nome do campo (em minúsculas)
      e o valor é a mensagem de erro.
    properties:
      errors:
        additionalProperties:
          type: string
        description: Mapa de erros de validação, onde a chave é o nome do campo (em
          minúsculas) e o valor é a mensagem de erro
        type: object
    type: object
  src_core_city_controller.ErrorMessageResponse:
    properties:
      message:
        type: string
    type: object
  src_core_city_controller.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API do Revit integrada com Swagger via Fiber.
  title: Revit API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Realiza login do usuário e retorna tokens de acesso e refresh
      parameters:
      - description: Credenciais de login (email ou nickname + senha + tipo de identificador)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/input.LoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens de autenticação
          schema:
            $ref: '#/definitions/response.AuthTokensResponse'
        "400":
          description: 'Erro na validação dos dados. Possíveis campos: identifier,
            password, identifierType'
          schema:
            $ref: '#/definitions/src_core_auth_controller.ValidationErrorResponse'
        "401":
          description: Credenciais inválidas
          schema:
            $ref: '#/definitions/src_core_auth_controller.ErrorMessageResponse'
      summary: Autenticar usuário
      tags:
      - Auth
  /api/auth/password:
    patch:
      consumes:
      - application/json
      description: Redefine a senha do usuário usando o token de redefinição
      parameters:
      - description: Token de redefinição e nova senha
        in: body
        name: resetData
        required: true
        schema:
          $ref: '#/definitions/input.ResetPassword'
      produces:
      - application/json
      responses:
        "204":
          description: Senha redefinida com sucesso
        "400":
          description: 'Erro na validação dos dados. Possíveis campos: resettoken,
            newpassword'
          schema:
            $ref: '#/definitions/src_core_auth_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_auth_controller.ErrorMessageResponse'
      summary: Redefinir senha
      tags:
      - Auth
    post:
      consumes:
      - application/json
      description: Envia email com token para redefinição de senha
      parameters:
      - description: Email ou nickname do usuário
        in: body
        name: identifier
        required: true
        schema:
          $ref: '#/definitions/input.Identifier'
      produces:
      - application/json
      responses:
        "204":
          description: Email enviado com sucesso
        "400":
          description: 'Erro na validação dos dados. Possíveis campos: identifier,
            identifierType'
          schema:
            $ref: '#/definitions/src_core_auth_controller.ValidationErrorResponse'
        "404":
          description: Usuário não encontrado
          schema:
            $ref: '#/definitions/src_core_auth_controller.ErrorMessageResponse'
      summary: Solicitar redefinição de senha
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Gera novos access token e refresh token, usando o refresh token
        atual
      parameters:
      - description: Bearer {refreshToken}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Novos tokens de autenticação
          schema:
            $ref: '#/definitions/response.AuthTokensResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_auth_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Atualizar tokens de autenticação
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma nova conta de usuário no sistema
      parameters:
      - description: Dados do usuário para registro
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/input.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Usuário criado com sucesso
          schema:
            $ref: '#/definitions/response.UserCreatedResponse'
        "400":
          description: 'Erro na validação dos dados. Possíveis campos: name, nickname,
            email, password, birthdate'
          schema:
            $ref: '#/definitions/src_core_auth_controller.ValidationErrorResponse'
        "409":
          description: Alguns dos campos únicos já estão em uso
          schema:
            $ref: '#/definitions/controller.ConflictErrorResponse'
      summary: Registrar novo usuário
      tags:
      - Auth
  /api/city:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de cidades filtradas por nome, com suporte a
        paginação
      parameters:
      - description: Nome da cidade para busca
        in: query
        name: name
        required: true
        type: string
      - description: Número da página (opcional)
        in: query
        name: page
        type: integer
      - description: Limite de resultados por página (opcional)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de cidades
          schema:
            items:
              $ref: '#/definitions/response.GetCityReponse'
            type: array
        "400":
          description: 'Erro na validação dos parâmetros. Possíveis campos: name,
            page, limit'
          schema:
            $ref: '#/definitions/src_core_city_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_city_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar cidades
      tags:
      - Cities
  /api/city/nearby:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de cidades próximas a uma localização geográfica
        (latitude e longitude)
      parameters:
      - description: Latitude da localização (-90 a 90)
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude da localização (-180 a 180)
        in: query
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Lista de cidades próximas
          schema:
            items:
              $ref: '#/definitions/response.GetCityReponse'
            type: array
        "400":
          description: 'Erro na validação dos parâmetros. Possíveis campos: latitude,
            longitude'
          schema:
            $ref: '#/definitions/src_core_city_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_city_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar cidades próximas
      tags:
      - Cities
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer " seguido do token JWT. Exemplo: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"